requirements.txt
requests>=2.31.0
python-dotenv>=1.0.0
pygame>=2.5.0
